+++ mrv2/lib/mrvApp/mrvSettingsObject.h
@@ -23,6 +23,7 @@ namespace mrv
         const char* kPenColorR = "Annotations/Pen Color R";
         const char* kPenColorG = "Annotations/Pen Color G";
         const char* kPenColorB = "Annotations/Pen Color B";
+        const char* kPenColorA = "Annotations/Pen Color A";
         const char* kPenSize = "Annotations/Pen Size";
         const char* kFontSize = "Annotations/Font Size";
         const char* kGhostPrevious = "Annotations/Ghost Previous";